requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.5
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-mysql==2.2.2
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
Markdown==2.6.11
mysqlclient==1.3.12
openpyxl==2.5.3
parso==0.2.1
pickleshare==0.7.4
Pillow==5.0.0
pipenv==2018.7.1
prompt-toolkit==1.0.15
psutil==5.4.6
psycopg2==2.7.5
py==1.5.2
Pygments==2.2.0
PyJWT==1.6.4
PyPDF2==1.26.0
pytest==3.2.5
pytz==2018.4
qrcode==5.3
reportlab==3.4.0
requests==2.19.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
whitenoise==3.3.1
workbook==1.1
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
